ThyroidClassificationProject/ai-service-3/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
python-multipart>=0.0.6
